+++ Typefast.Server/Data/DTOs/BlockRequest.cs
@@ -3,7 +3,7 @@ namespace Typefast.Server.Data.DTOs
     public class BlockRequest
     {
         public required int IdPer { get; set; }
-        public required int blUntil { get; set; }
+        public required long blUntil { get; set; }
         public required string blReason { get; set; }
     }
 }
